login module

+++ src/main/java/com/anyu/system/controller/EmployeeController.java
@@ -21,14 +21,8 @@ public class EmployeeController {
     @Resource
     private EmployeeService employeeService;
 
-    /**
-     * 通过主键查询单条数据
-     *
-     * @param id 主键
-     * @return 单条数据
-     */
     @GetMapping("selectOne")
-    public Employee selectOne(Long id) {
+    public Employee selectOne(Integer id) {
         return this.employeeService.getById(id);
     }
